Face_Recognition/requirements.txt
Flask==2.3.3
# tensorflow==2.18.0
opencv-python==4.11.0.86
numpy>=1.26.0,<2.1.0
requests==2.32.3
python-dotenv==1.0.0

geographiclib==2.1
geopy==2.4.1
pillow==12.0.0